viridis_pal {viridis}	R Documentation

Viridis Color Palettes

Description

A wrapper function around viridis to turn it into a palette function compatible with discrete_scale.

Usage

viridis_pal(alpha = 1, begin = 0, end = 1, direction = 1, option = "D")

Arguments

alpha
The alpha transparency, a number in [0,1], see argument alpha in hsv.

begin
The (corrected) hue in [0,1] at which the color map begins.

end
The (corrected) hue in [0,1] at which the color map ends.

direction
Sets the order of colors in the scale. If 1, the default, colors are ordered from darkest to lightest. If -1, the order of colors is reversed.

option
A character string indicating the color map option to use. Eight options are available:

"magma" (or "A")

"inferno" (or "B")

"plasma" (or "C")

"viridis" (or "D")

"cividis" (or "E")

"rocket" (or "F")

"mako" (or "G")

"turbo" (or "H")

Details

See viridis and viridis.map for more information on the color palettes.

Author(s)

Bob Rudis: bob@rud.is

Simon Garnier: garnier@njit.edu

Examples

library(scales)
show_col(viridis_pal()(12))

[Package viridis version 0.6.5 Index]